fix prun

+++ pruning/prune.py
@@ -112,9 +112,9 @@ def prune_vgg16_conv_layer(model, layer_index, filter_index):
 		new_weights = new_linear_layer.weight.data.cpu().numpy()	 	
 
 		new_weights[:, : filter_index * params_per_input_channel] = \
-			old_weights[:, : filter_index * params_per_input_channel]
-		new_weights[:, filter_index * params_per_input_channel :] = \
-			old_weights[:, (filter_index + 1) * params_per_input_channel :]
+			old_weights[:, : int(filter_index * params_per_input_channel)]
+		new_weights[:, int(filter_index * params_per_input_channel ):] = \
+			old_weights[:, int((filter_index + 1) * params_per_input_channel) :]
 
 		new_linear_layer.bias.data = old_linear_layer.bias.data
